🚚 Move Parser classes to different directories for each features
