+++ src/internal.ts
@@ -3,19 +3,19 @@
  * @see https://medium.com/visual-development/how-to-fix-nasty-circular-dependency-issues-once-and-for-all-in-javascript-typescript-a04c987cf0de
  */
 
-export * from './parser';
-export * from './parser/internal/converter';
-export * from './parser/internal/any-or-more';
-export * from './parser/internal/value-converter';
-export * from './parser/action';
-export * from './parser/any-character';
-export * from './parser/character-class';
-export * from './parser/customizable';
-export * from './parser/literal-string';
-export * from './parser/matched-text';
-export * from './parser/one-or-more';
-export * from './parser/optional';
-export * from './parser/times';
-export * from './parser/value';
-export * from './parser/zero-or-more';
+export * from './parser/internal';
+export * from './parser/primary/any-character';
+export * from './parser/primary/character-class';
+export * from './parser/primary/customizable';
+export * from './parser/primary/literal-string';
+export * from './parser/modifier/internal/converter';
+export * from './parser/modifier/internal/any-or-more';
+export * from './parser/modifier/internal/value-converter';
+export * from './parser/modifier/action';
+export * from './parser/modifier/matched-text';
+export * from './parser/modifier/one-or-more';
+export * from './parser/modifier/optional';
+export * from './parser/modifier/times';
+export * from './parser/modifier/value';
+export * from './parser/modifier/zero-or-more';
 export * from './parser-generator';
