+++ src/parser/modifier/internal/converter.ts
@@ -1,4 +1,4 @@
-import { Parser } from '../../internal';
+import { Parser } from '../../../internal';
 
 export abstract class ConverterParser<
     TPrevResult,
